clarify difference between extension and linking semantics, as well as respective syntax
Extending another struct is done based on the current value of the
original struct, any changes to it done later on in the parse will not
be reflected. In contrast, links are followed at runtime, not during parse.

Since the syntax for the two types of links (immediate and lazy) are
now different, in theory we could add lazy links to extending and
immediate links for attribute values.