Fixing previous broken test cases by expanding flattened paths during parsing.
Basically this fix addresses the case where you have the key a.b.c.d.e.f: "some value"
the parser creates the prototypes a, b, c, d and e before attempting to populate e with f.
Since it is possible that none of the parents of f were defined outside of the key itself.