Update test suite to show struct comparison is broken.
I noticed that while attempting to write a new test that it didn't catch
an error that I had made. I unified the pass/fail cases so the fail
tests can also do struct comparison as a basis for testing and sure
enough two different structs are failing to fail that test.

I wonder if there are any tests that are falsely passing due to this.