Consider attribute order in Struct equality tests.
Previous versions of 0.3.x used the default unordered dict comparison
when testing equality. So even though Structs are ordered mapping
objects order didn't matter in comparisons. This lead to a surprise in
the unit tests, for a long while I couldn't figure out how the @map unit
test was actually passing because the ordering was wrong.

This is an incompatible change so it is not suitable for 0.3.x.