Always wrap values inside List and Struct objects in Nodes.
This will allow future code to handle most object identically in a
polymorphic manner without so many isinstance crazyness.

To help support this Leaf objects now passes all comparisons directly
through to the value it is wrapping. This allows comparisons of coil
structures to built-in python types to just work as before.

Also, it turns out that subclassing list is a little obnoxious in that
none of the other methods go through the custom getitem/setitem
implementation so I must re-implement most of the functions involved in
getting and setting things.